+++ src/ui/main/FileScannerDialog.cpp
@@ -34,24 +34,23 @@ FileScannerDialog::FileScannerDialog(QWidget* parent) : QDialog(parent), ui(new 
     connect(manager->tvShowFileSearcher(),  &TvShowFileSearcher::progress,  this, &FileScannerDialog::onProgress);
     connect(manager->musicFileSearcher(),   &MusicFileSearcher::progress,   this, &FileScannerDialog::onProgress);
 
-    connect(manager->movieFileSearcher(),   &MovieFileSearcher::currentDir, this, [this](QString dir){
+    connect(manager->movieFileSearcher(),   &MovieFileSearcher::progressText, this, [this](QString dir){
         ui->currentDir->setText(dir);
-        // The next line is disabled on purpose, see https://github.com/Komet/MediaElch/issues/1315
-        // The issue seems to be an event in Qt's event queue that is executed after the object
-        // was deleted by deleteLater().
+        // Do not enable the following line. The movie file searcher is currently the only
+        // one that uses multithreading which means there is no need for this call.
         // QApplication::processEvents();
     });
     connect(manager->concertFileSearcher(), &ConcertFileSearcher::currentDir, this, &FileScannerDialog::onCurrentDir);
     connect(manager->tvShowFileSearcher(),  &TvShowFileSearcher::currentDir,  this, &FileScannerDialog::onCurrentDir);
     connect(manager->musicFileSearcher(),   &MusicFileSearcher::currentDir,   this, &FileScannerDialog::onCurrentDir);
 
-    connect(manager->movieFileSearcher(),   &MovieFileSearcher::searchStarted,   ui->status, &QLabel::setText);
+    connect(manager->movieFileSearcher(),   &MovieFileSearcher::statusChanged,   ui->status, &QLabel::setText);
     connect(manager->tvShowFileSearcher(),  &TvShowFileSearcher::searchStarted,  ui->status, &QLabel::setText);
     connect(manager->concertFileSearcher(), &ConcertFileSearcher::searchStarted, ui->status, &QLabel::setText);
     connect(manager->musicFileSearcher(),   &MusicFileSearcher::searchStarted,   ui->status, &QLabel::setText);
     // clang-format on
 
-    connect(manager->movieFileSearcher(), &MovieFileSearcher::moviesLoaded, [this]() {
+    connect(manager->movieFileSearcher(), &MovieFileSearcher::finished, [this]() {
         if (m_reloadType != ReloadType::All) {
             accept();
         } else {
@@ -135,39 +134,43 @@ void FileScannerDialog::setReloadType(ReloadType type)
 }
 
 /**
- * \brief Rejecting should not be possible
+ * \brief Rejected, e.g. by pressing ESC
  */
 void FileScannerDialog::reject()
 {
+    // Note: We use "singleShot" simply because the same is done for reload().
+    //       If we call it directly, abort() may be called _before_ reload().
+
     if (m_reloadType == ReloadType::Movies || m_reloadType == ReloadType::All) {
-        Manager::instance()->movieFileSearcher()->abort();
-        Manager::instance()->movieModel()->clear();
+        QTimer::singleShot(0, this, []() { Manager::instance()->movieFileSearcher()->abort(); });
+        // Don't clear when aborted. All movies that are shown, exist in the database.
     }
     if (m_reloadType == ReloadType::TvShows || m_reloadType == ReloadType::All) {
-        Manager::instance()->tvShowFileSearcher()->abort();
-        Manager::instance()->tvShowModel()->clear();
-        Manager::instance()->tvShowFilesWidget()->renewModel();
+        QTimer::singleShot(0, this, []() {
+            Manager::instance()->tvShowFileSearcher()->abort();
+            Manager::instance()->tvShowModel()->clear();
+            Manager::instance()->tvShowFilesWidget()->renewModel();
+        });
     }
     if (m_reloadType == ReloadType::Concerts || m_reloadType == ReloadType::All) {
-        Manager::instance()->concertFileSearcher()->abort();
-        Manager::instance()->concertModel()->clear();
+        QTimer::singleShot(0, this, []() {
+            Manager::instance()->concertFileSearcher()->abort();
+            Manager::instance()->concertModel()->clear();
+        });
     }
     if (m_reloadType == ReloadType::Music || m_reloadType == ReloadType::All) {
-        Manager::instance()->musicFileSearcher()->abort();
-        Manager::instance()->musicModel()->clear();
+        QTimer::singleShot(0, this, []() {
+            Manager::instance()->musicFileSearcher()->abort();
+            Manager::instance()->musicModel()->clear();
+        });
     }
 
     QDialog::reject();
 }
 
-/**
- * \brief Starts the movie file searcher
- */
 void FileScannerDialog::onStartMovieScanner()
 {
     ui->progressBar->setValue(0);
-    Manager::instance()->movieModel()->clear();
-    QApplication::processEvents();
     if (m_forceReload) {
         QTimer::singleShot(0, this, &FileScannerDialog::onStartMovieScannerForce);
     } else {
@@ -190,7 +193,6 @@ void FileScannerDialog::onStartTvShowScanner()
 {
     ui->progressBar->setValue(0);
     Manager::instance()->tvShowModel()->clear();
-    QApplication::processEvents();
     if (m_forceReload) {
         QTimer::singleShot(0, this, &FileScannerDialog::onStartTvShowScannerForce);
     } else {
@@ -220,7 +222,6 @@ void FileScannerDialog::onStartConcertScanner()
 {
     ui->progressBar->setValue(0);
     Manager::instance()->concertModel()->clear();
-    QApplication::processEvents();
     if (m_forceReload) {
         QTimer::singleShot(0, this, &FileScannerDialog::onStartConcertScannerForce);
     } else {
@@ -242,7 +243,6 @@ void FileScannerDialog::onStartMusicScanner()
 {
     ui->progressBar->setValue(0);
     Manager::instance()->musicModel()->clear();
-    QApplication::processEvents();
     if (m_forceReload) {
         QTimer::singleShot(0, this, &FileScannerDialog::onStartMusicScannerForce);
     } else {
@@ -278,6 +278,7 @@ void FileScannerDialog::onProgress(int current, int max)
 void FileScannerDialog::onCurrentDir(QString dir)
 {
     ui->currentDir->setText(dir);
+    // TODO: Remove once all file searchers run in another thread.
     QApplication::processEvents();
 }